compiles now

+++ packages/exedao-sol/contracts/Extendable.sol
@@ -8,9 +8,9 @@ contract Extendable is Permissioned {
   using ExeLib for address;
   using ExeLib for bytes;
   ExeLib.Extension[] public extensions;
-  mapping(bytes4 => uint) public extensionFor;
+  mapping(bytes4 => uint256) public extensionFor;
 
-  event ExtensionAdded(uint extensionIndex, bytes32 metaHash);
+  event ExtensionAdded(uint256 extensionIndex, bytes32 metaHash);
 
   constructor(
     uint64 shares, uint64 _proposalDuration,
@@ -38,18 +38,18 @@ contract Extendable is Permissioned {
         extension.extensionAddress = extension.bytecode.deploy();
         delete extension.bytecode;
       }
-      uint index = extensions.length;
+      uint256 index = extensions.length;
       extensions.push(extension);
       bytes4[] memory funcSigs = extension.functionSignatures;
-      for (uint i = 0; i < funcSigs.length; i++) extensionFor[funcSigs[i]] = index;
+      for (uint256 i = 0; i < funcSigs.length; i++) extensionFor[funcSigs[i]] = index;
       emit ExtensionAdded(index, extension.metaHash);
     }
   }
 
-  function removeExtension(uint extIndex) external {
+  function removeExtension(uint256 extIndex) external {
     if (voteAndContinue()) {
       ExeLib.Extension memory ext = extensions[extIndex];
-      for (uint i = 0; i < ext.functionSignatures.length; i++) {
+      for (uint256 i = 0; i < ext.functionSignatures.length; i++) {
         bytes4 funcSig = ext.functionSignatures[i];
         delete extensionFor[funcSig];
         if (approvalRequirements[funcSig] != 0) delete approvalRequirements[funcSig];
